Rename BrewUI targets to Brew for naming clarity
