+++ .github/workflows/ui_smoke.yml
@@ -20,25 +20,44 @@ jobs:
     name: Run UI smoke test
     runs-on: macos-26
     timeout-minutes: 40
+    env:
+      CI_APPLE_DEVELOPER_TEAM_ID: ${{ secrets.CI_APPLE_DEVELOPER_TEAM_ID }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
 
       - name: Resolve package dependencies
-        run: xcodebuild -resolvePackageDependencies -project BrewUI.xcodeproj
+        run: xcodebuild -resolvePackageDependencies -project Brew.xcodeproj
 
       - name: Run UI smoke test
         run: |
           set -euo pipefail
           TEST_FILTER="${{ github.event.inputs.test_filter }}"
           if [[ -z "$TEST_FILTER" ]]; then
-            TEST_FILTER="BrewUIUITests/BrewUIUITests/testExample"
+            TEST_FILTER="BrewUITests/BrewUITests/testExample"
+          fi
+
+          # UI test bundles must be signed. In CI, use ad-hoc signing by default
+          # (no certificate required). If a CI development team ID secret is
+          # configured, prefer automatic Apple Development signing instead.
+          SIGNING_ARGS=(
+            CODE_SIGN_STYLE=Manual
+            CODE_SIGN_IDENTITY=-
+            CODE_SIGNING_REQUIRED=NO
+          )
+          if [[ -n "${CI_APPLE_DEVELOPER_TEAM_ID:-}" ]]; then
+            SIGNING_ARGS=(
+              CODE_SIGN_STYLE=Automatic
+              CODE_SIGN_IDENTITY="Apple Development"
+              DEVELOPMENT_TEAM="$CI_APPLE_DEVELOPER_TEAM_ID"
+            )
           fi
 
           xcodebuild test \
-            -project BrewUI.xcodeproj \
-            -scheme BrewUI \
+            -project Brew.xcodeproj \
+            -scheme Brew \
             -destination "platform=macOS" \
             -only-testing "$TEST_FILTER" \
             -skipPackagePluginValidation \
-            -skipMacroValidation
+            -skipMacroValidation \
+            "${SIGNING_ARGS[@]}"
