+++ .swiftlint.yml
@@ -1,7 +1,7 @@
 included:
-  - BrewUI
+  - Brew
+  - BrewTests
   - BrewUITests
-  - BrewUIUITests
 
 excluded:
   - .build
