+++ BrewTests/BrewTests.swift
@@ -1,0 +1,15 @@
+//
+//  BrewTests.swift
+//  BrewTests
+//
+//  Created by Graeme Arthur on 6/3/2026.
+//
+
+@testable import Brew
+import Testing
+
+struct BrewTests {
+    @Test func example() {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+}
